requirements.txt
annotated-types~=0.7.0
azure-cognitiveservices-speech~=1.41.1
cachetools~=5.5.0
certifi~=2024.8.30
charset-normalizer~=3.4.0
click~=8.1.7
cloup~=3.0.5
Cython~=3.0.11
decorator~=5.1.1
glcontext~=3.0.0
google-ai-generativelanguage~=0.6.10
google-api-core~=2.22.0
google-api-python-client~=2.151.0
google-auth~=2.35.0
google-auth-httplib2~=0.2.0
google-generativeai~=0.8.3
googleapis-common-protos~=1.65.0
grpcio~=1.67.1
grpcio-status~=1.67.1
gTTS~=2.5.3
httplib2~=0.22.0
idna~=3.10
isosurfaces~=0.1.2
manim~=0.18.1
manim-voiceover~=0.3.7
ManimPango~=0.6.0 # sudo apt-get install libsdl-pango-dev    if you dont have pangocairo
mapbox_earcut~=1.0.2
markdown-it-py~=3.0.0
mdurl~=0.1.2
moderngl~=5.12.0
multipledispatch~=1.0.0
mutagen~=1.47.0
networkx~=3.4.2
numpy~=2.2.2
pillow
proto-plus~=1.25.0
protobuf~=5.28.3
pyasn1~=0.6.1
pyasn1_modules~=0.4.1
PyAudio~=0.2.14 #required brew install portaudio for mac
pycairo~=1.27.0
pydantic~=2.9.2
pydantic_core~=2.23.4
pydub~=0.25.1
pyglet~=2.0.18
Pygments~=2.18.0
#pyobjc-core~=10.3.1 # only for mac
#pyobjc-framework-Cocoa~=10.3.1 # only for mac
pyparsing~=3.2.0
pyrr~=0.10.3
python-dotenv~=0.21.1
python-slugify~=8.0.4
requests~=2.32.3
rich~=13.9.3
rsa~=4.9
scipy~=1.14.1
screeninfo~=0.8.1
skia-pathops~=0.8.0.post2
sox~=1.5.0
srt~=3.5.3
svgelements~=1.9.6
text-unidecode~=1.3
tqdm~=4.66.5
typing_extensions~=4.12.2
uritemplate~=4.1.1
urllib3~=2.2.3
watchdog~=5.0.3
inquirer
openai~=1.61.0
tiktoken~=0.8.0
timm
sentencepiece
transformers
litellm~=1.60.5
pysrt
moviepy~=2.1.2
yt-dlp
imageio_ffmpeg~=0.5.1
langchain~=0.3.14
langchain_community~=0.3.14
SpeechRecognition~=3.14.1
boto3~=1.36.9
manim-physics~=0.4.0
manim-ml~=0.0.24
manim-chemistry~=0.4.4
manim-dsa~=0.2.0
manim-circuit~=0.0.3
langfuse~=2.58.1
chromadb~=0.6.3
google-cloud-aiplatform~=1.79.0
cairosvg
pylatexenc~=2.10
ffmpeg-python~=0.2.0
kokoro-onnx[gpu] # if you have a GPU, otherwise kokoro-onnx
soundfile~=0.13.1
krippendorff~=0.8.1
statsmodels~=0.14.4
opencv-python~=4.11.0
gradio~=4.29.0
